+++ src/entities/mypage/model/types.ts
